+++ profiles/nlgov.js
@@ -7,17 +7,17 @@ const modules = [
   import("../src/core/location-hash.js"),
   import("../src/core/l10n.js"),
   import("../src/core/style.js"),
-  import("../src/logius/style.js"), // done for now, still some wip
+  import("../src/nlgov/style.js"), // done for now, still some wip
   import("../src/core/github.js"),
   import("../src/core/data-include.js"),
   import("../src/core/markdown.js"),
   import("../src/core/reindent.js"), // nothing changed but this module is in geonovum profile at line 3
-  import("../src/logius/title.js"),
+  import("../src/nlgov/title.js"),
   import("../src/core/title.js"),
   import("../src/w3c/level.js"),
   import("../src/w3c/group.js"),
-  import("../src/logius/missing-config-warner.js"),
-  import("../src/logius/headers.js"),
+  import("../src/nlgov/missing-config-warner.js"),
+  import("../src/nlgov/headers.js"),
   import("../src/w3c/abstract.js"),
   import("../src/core/data-transform.js"),
   import("../src/core/data-abbr.js"),
@@ -26,7 +26,7 @@ const modules = [
   import("../src/core/dfn.js"),
   import("../src/core/pluralize.js"),
   import("../src/core/examples.js"),
-  import("../src/logius/issues-notes.js"),
+  import("../src/nlgov/issues-notes.js"),
   import("../src/core/best-practices.js"),
   import("../src/core/figures.js"),
   import("../src/core/biblio.js"),
@@ -54,8 +54,8 @@ const modules = [
   import("../src/core/algorithms.js"),
   import("../src/core/anchor-expander.js"),
   import("../src/core/custom-elements/index.js"),
-  import("../src/logius/label.js"),
-  import("../src/logius/override.js"),
+  import("../src/nlgov/label.js"),
+  import("../src/nlgov/override.js"),
   /* Linters must be the last thing to run */
   import("../src/core/linter-rules/check-charset.js"),
   import("../src/core/linter-rules/check-punctuation.js"),
